Ajout d'un Timer sur Clients afin que le groupe s'en aille si il attends trop longtemps

+++ ProjetProgrammationSystème/Salle/Controller/StaffInterface.cs
@@ -8,7 +8,7 @@ namespace Salle.Controller
 {
     public interface StaffInterface
     {
-        int Schedule
+        long Schedule
         {
             get;
             set;
